+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## 1.3.2 (`1030270`)
 
+- **New:** Button to clear email logs (#90)
 - **Change:** Options, Permissions and Style Properties are now grouped in the output rather than displaying group inline (#88)
 - **Change:** Increase mails per-page to 100 (#89)
 - **Fix:** Fix newline printing literally in the resulting Markdown code (#88)